pyproject.toml
[tool.pytest.ini_options]
minversion = "4.0"
addopts = " --cov=asset_manager"

[project]
name = "asset_manager"
version = "0.1.0"
description = ""
authors = [{name = "Ethan Swan", email = "ethanpswan@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altair>=5.4.0",
    "google-api-python-client>=2.66.0",
    "google-auth-oauthlib>=0.7.1",
    "psycopg[binary]>=3.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "yoyo-migrations>=8.0.0",
    # Legacy deps - needed for S3 migration script, can remove after migration
    "boto3>=1.26.13",
    "pandas>=1.5.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.2.0",
    "pytest-cov>=4.0.0",
    "mypy>=0.991",
    "ipython>=8.6.0",
    "hypothesis>=6.57.1",
    "google-api-python-client-stubs>=1.13.0",
    "boto3-stubs[essential]>=1.26.13",
    "black>=22.10.0",
    "jupyterlab>=3.5.0",
    "types-setuptools>=65.6.0.1",
    "pandas-stubs>=1.5.2.221124",
    "testcontainers[postgres]>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.12.9",
]
